requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
